Profile shows current users information
You can now edit user fields for the currently logged in user.
This syncs automatically to the db after each edit.
Also, the login button changes to a log out button and a profile button
when logged in. A couple minor changes in the backend had to be made
to allow for updating the user fields, instead of creating a new user.
In addition, some minor changes were made to the css files.